completed the getSingleIssue controller
