+++ src/middleware/isAuthenticated.js
@@ -1,0 +1,8 @@
+const isAuthenticated = (req, res, next) => {
+    if (req.isAuthenticated()) {
+        return next();
+    }
+    res.redirect('/login');
+};
+
+module.exports = { isAuthenticated };
