+++ src/models/comment.model.js
@@ -1,11 +1,7 @@
 const mongoose = require('mongoose');
 
 const CommentSchema = new mongoose.Schema({
-  content: {
-    type: String,
-    required: true,
-    trim: true
-  },
+  content: {type: String, required: true, trim: true},
   author: {
     type: mongoose.Schema.Types.ObjectId,
     ref: 'User',
